Use a custom post status to track post visibility
This custom post status is registered as private so that query
parsing happens in the same way that private posts do, but in a
way that does not mirror the behavior of private posts exactly.

If a restricted page view loads, we log a redirect flag that helps
to decide when to redirect a user if a 404 view has been reached.