Remove the custom post status and use private instead
Really what we're trying to do is alter the behavior of the post
submit actions box rather than actually change the status of a
post. "Private" works for what we're trying to do and already does
a good job of hiding the post from the world by default in all
views. We use post meta to determine what subset of logged in users
can see that private post.